whisperx_spark/requirements/requirements_cuda.txt
whisperx>=3.1.1
faster-whisper>=1.1.0
ctranslate2>=4.4.0
ffmpeg-python>=0.2.0
audioread>=3.0.1
soundfile>=0.12.1
librosa>=0.10.2
av>=14.0.1
gradio>=5.9.0
gradio_client>=1.5.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
transformers>=4.40.0
huggingface-hub>=0.27.0
tokenizers>=0.21.0
safetensors>=0.4.0
onnxruntime>=1.20.0
accelerate>=1.2.0
asteroid-filterbanks>=0.4.0
speechbrain>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
coloredlogs>=15.0.0
colorlog>=6.9.0
tqdm>=4.66.0
pyyaml>=6.0.0
numpy>=1.26.0,<2.0.0
psutil>=5.9.0
